+++ src/components/components/Drawer/Styled.ts
@@ -4,13 +4,14 @@ import { DrawerProps } from "./types";
 export const DrawerContainer = styled.div`
   position: fixed;
   padding: ${props => `${props.theme.space[0]}px`};
-  width: 0;
+  width: auto;
+  height: auto;
   top: 0;
   z-index: 1000;
 `;
 DrawerContainer.displayName = "DrawerContainer";
 
-export const DrawerPanel = styled.div<DrawerProps>`
+export const DrawerPanel = styled.div`
   position: fixed;
   top: 0;
   left: 0;
@@ -18,17 +19,15 @@ export const DrawerPanel = styled.div<DrawerProps>`
   width: auto;
   max-width: 280px;
   background-color: #fff;
-  transform: ${props => !props.isOpen && `translateX(-280px)`};
+  z-index: 1;
+  transform: translateX(-280px);
   transition: transform 0.3s ease;
   will-change: transform;
-  z-index: 1;
 `;
+
 DrawerPanel.displayName = "DrawerPanel";
 
 export const Overlay = styled.div<DrawerProps>`
-  opacity: ${props => (props.isOpen ? 1 : 0)};
-  visibility: ${props => (props.isOpen ? "visible" : "hidden")};
-  transition: opacity 0.1s ease;
   background: rgba(0, 18, 35, 0.75);
   position: fixed;
   bottom: 0;
@@ -36,5 +35,8 @@ export const Overlay = styled.div<DrawerProps>`
   right: 0;
   top: 0;
   z-index: 0;
+  opacity: ${props => (props.isOpen ? 1 : 0)};
+  visibility: ${props => (props.isOpen ? "visible" : "hidden")};
+  transition: opacity 0.5s ease;
 `;
 Overlay.displayName = "DrawerOverlay";